updatedng10/11

+++ ANGULAR/lp1ang18/src/app/list-users/list-users.component.ts
@@ -13,6 +13,8 @@ import { Router } from '@angular/router';
 export class ListUsersComponent implements OnInit{
   users:User[];
   userService:UserService;
+  searchinp:number;
+  searchedusers:User[];
   constructor(userService:UserService,private router:Router){
     this.userService=userService;
   }
@@ -20,9 +22,25 @@ ngOnInit(): void {
   this.userService.getUser().subscribe((userdata)=>{this.users=userdata;})
 }
 
-deleteUser(id:number){
-  this.userService.deleteUser(id).subscribe(()=>
-  this.users=this.users.filter(user=>user.id!=id))
+deleteUser(id:number){  
+  if(id!==undefined){
+    this.userService.deleteUser(id).subscribe(()=>
+      this.users=this.users.filter(user=>user.id!=id))
+  }
 }
 
+updateUser(id:number|undefined):void{
+  if(id!==undefined){
+    this.router.navigate(['update',id]);
+  }
+}
+searchUser(){
+  this.userService.getUser().subscribe((emp)=>{
+    this.searchedusers=emp.filter(emp=>emp.id==this.searchinp)
+  })
+}
+logut(){
+  sessionStorage.setItem("loggedIn","no");
+  this.router.navigate(["/"]);
+}
 }
